+++ lib/mocks.dart
@@ -34,8 +34,10 @@ final List<Sight> mocks = [
 ];
 
 List<String> history = [];
-List<Sight> visSight = [];
-List<Sight> visitedSight = [];
+List<Sight> visSight =
+    mocks.where((sight) => sight.wantToVisit == true).toList();
+List<Sight> visitedSight =
+    mocks.where((sight) => sight.visited == true).toList();
 List<String> newSightsMocksPhotosList = [
   'https://kuda-mo.ru/uploads/9565f99c9a90ddbbbe2a671c548c28a2.jpeg',
   'https://ic.pics.livejournal.com/alik_morozov/68038780/1002703/1002703_original.jpg',
