+++ lib/ui/screen/SightSearchScreen.dart
@@ -4,7 +4,7 @@ import 'package:places/domain/sight.dart';
 import 'package:places/res/Strings.dart';
 import 'package:places/res/colors.dart';
 import '../../mocks.dart';
-import 'Sight_details.dart';
+import 'sight_details.dart';
 
 class SearchScreen extends StatefulWidget {
   SearchScreen({Key key}) : super(key: key);
