+++ lib/ui/screen/VisitingScreen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
-import 'package:places/domain/sight.dart';
+import 'package:places/components/sightVisited.dart';
+import 'package:places/components/sightWantVisit.dart';
 import 'package:places/res/Strings.dart';
 import 'package:places/res/colors.dart';
 import '../../mocks.dart';
-import 'Sight_details.dart';
 
 class FavoriteScreen extends StatefulWidget {
   FavoriteScreen({Key key}) : super(key: key);
@@ -14,8 +14,6 @@ class FavoriteScreen extends StatefulWidget {
 }
 
 class _FavoriteScreenState extends State<FavoriteScreen> {
-  List wantVisit = [];
-  List visitedList = [];
   @override
   Widget build(BuildContext context) {
     return DefaultTabController(
@@ -69,20 +67,14 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
   }
 
   void _onChange() {
-    wantVisit = mocks.where((sight) => sight.wantToVisit == true).toList();
-    visitedList = mocks.where((sight) => sight.visited == true).toList();
     setState(() {
-      visSight = wantVisit;
-      visitedSight = visitedList;
+      visSight = mocks.where((sight) => sight.wantToVisit == true).toList();
+      visitedSight = mocks.where((sight) => sight.visited == true).toList();
     });
   }
 
   Widget sightWantToVisitList() {
-    setState(() {
-      wantVisit = mocks.where((sight) => sight.wantToVisit == true).toList();
-      visSight = wantVisit;
-    });
-    if (wantVisit.length > 0) {
+    if (visSight.length > 0) {
       return SingleChildScrollView(
         child: ListView.builder(
           shrinkWrap: true,
@@ -95,8 +87,7 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
               setState(() {});
             },
             onDelete: () {
-              wantVisit.remove(visSight[index]);
-              visitedList.add(visSight[index]);
+              visitedSight.add(visSight[index]);
               _onChange();
             },
           ),
@@ -107,10 +98,6 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
   }
 
   Widget sightVisited() {
-    setState(() {
-      visitedList = mocks.where((sight) => sight.visited == true).toList();
-      visitedSight = visitedList;
-    });
     if (visitedSight.length > 0) {
       return SingleChildScrollView(
         child: ListView.builder(
@@ -121,7 +108,7 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
             sight: visitedSight[index],
             key: ObjectKey(visitedSight[index]),
             onDelete: () {
-              visitedList.remove(visitedSight[index]);
+              visitedSight.remove(visitedSight[index]);
               _onChange();
             },
           ),
@@ -185,335 +172,6 @@ class _FavoriteScreenState extends State<FavoriteScreen> {
                   .copyWith(fontSize: 17, color: secondary2),
               textAlign: TextAlign.center),
         ],
-      ),
-    );
-  }
-}
-
-class FavoriteWishVisit extends StatelessWidget {
-  final Sight sight;
-  final VoidCallback onDelete;
-  final VoidCallback onFilterChange;
-
-  const FavoriteWishVisit(
-      {Key key, this.sight, this.onDelete, this.onFilterChange});
-  static const double Height = 120;
-
-  @override
-  Widget build(BuildContext context) {
-    return DragTarget<Sight>(
-      builder: (
-        BuildContext context,
-        List<Sight> candidateData,
-        List<dynamic> rejectedData,
-      ) {
-        return LongPressDraggable(
-          data: sight,
-          axis: Axis.vertical,
-          feedback: ConstrainedBox(
-              constraints:
-                  BoxConstraints(maxWidth: MediaQuery.of(context).size.width),
-              child: sightCard(context)),
-          child: sightCard(context),
-          childWhenDragging: const SizedBox.shrink(),
-        );
-      },
-      // onAccept: (newFilteredSight) {
-      //   final int oldIndex = visSight.indexOf(sight);
-      //   final int newIndex = visSight.indexOf(newFilteredSight);
-      //   visSight.removeAt(oldIndex);
-      //   visSight.insert(newIndex, sight);
-      // visSight[oldIndex] = newFilteredSight;
-      // visSight[newIndex] = sight;
-      //   onFilterChange();
-      // },
-    );
-  }
-
-  Widget sightCard(BuildContext context) {
-    return Material(
-      type: MaterialType.transparency,
-      child: InkWell(
-        onTap: () {
-          Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (context) => SightDetail(sight: sight)));
-        },
-        child: Container(
-          padding: EdgeInsets.only(left: 16, right: 16, top: 10, bottom: 30),
-          child: Column(
-            children: [
-              Stack(
-                children: [
-                  Container(
-                    child: ClipRRect(
-                      borderRadius: const BorderRadius.only(
-                        topLeft: const Radius.circular(15),
-                        topRight: const Radius.circular(15),
-                      ),
-                      child: Image.network(
-                        '${sight.url}',
-                        height: double.infinity,
-                        width: double.infinity,
-                        fit: BoxFit.cover,
-                        loadingBuilder: (context, child, loadingProgress) {
-                          if (loadingProgress == null) return child;
-
-                          return Center(
-                            child: CircularProgressIndicator(
-                              value: loadingProgress.expectedTotalBytes != null
-                                  ? loadingProgress.cumulativeBytesLoaded /
-                                      loadingProgress.expectedTotalBytes
-                                  : null,
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                    height: FavoriteWishVisit.Height,
-                    decoration: BoxDecoration(
-                      color: Theme.of(context).backgroundColor,
-                      borderRadius: const BorderRadius.only(
-                        topLeft: const Radius.circular(15),
-                        topRight: const Radius.circular(15),
-                      ),
-                    ),
-                  ),
-                  Positioned(
-                      top: 16,
-                      left: 16,
-                      child: Text(
-                        '${sight.titleType}',
-                        style: Theme.of(context).textTheme.bodyText2,
-                      )),
-                  Positioned(
-                    //!button share
-                    top: 1,
-                    right: 40,
-                    child: IconButton(
-                      icon: Icon(Icons.share, size: 28, color: iconColor),
-                      onPressed: () {
-                        print('button "share" pressed');
-                        onDelete();
-                      },
-                    ),
-                  ),
-                  Positioned(
-                    //!button close
-                    top: 0,
-                    right: 0,
-                    child: IconButton(
-                      icon: Icon(Icons.close, size: 30, color: iconColor),
-                      onPressed: () {
-                        sight.wantToVisit = !sight.wantToVisit;
-                        sight.visited = !sight.visited;
-                        print(' want to visit $sight.wantToVisit');
-                        onDelete();
-                      },
-                    ),
-                  ),
-                ],
-              ),
-              Container(
-                decoration: BoxDecoration(
-                  color: Theme.of(context).backgroundColor,
-                  borderRadius: const BorderRadius.only(
-                    bottomLeft: const Radius.circular(15),
-                    bottomRight: const Radius.circular(15),
-                  ),
-                ),
-                padding: const EdgeInsets.all(16),
-                width: double.infinity,
-                height: FavoriteWishVisit.Height,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      '${sight.name}',
-                      maxLines: 2,
-                      style: Theme.of(context).textTheme.headline5,
-                    ),
-                    SizedBox(
-                      height: 5,
-                    ),
-                    Text(
-                      'Запланировано на 12 окт.2020',
-                      style: Theme.of(context).textTheme.bodyText1,
-                    ),
-                    SizedBox(
-                      height: 5,
-                    ),
-                    Text(
-                      '${sight.workHours}',
-                      style: Theme.of(context).textTheme.subtitle2,
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class FavoriteVisited extends StatelessWidget {
-  final Sight sight;
-  final VoidCallback onDelete;
-  final VoidCallback onFilterChange;
-
-  const FavoriteVisited(
-      {Key key, this.sight, this.onDelete, this.onFilterChange});
-  static const double Height = 120;
-  @override
-  Widget build(BuildContext context) {
-    return DragTarget<Sight>(
-      builder: (
-        BuildContext context,
-        List<Sight> candidateData,
-        List<dynamic> rejectedData,
-      ) {
-        return LongPressDraggable(
-          data: sight,
-          axis: Axis.vertical,
-          feedback: ConstrainedBox(
-              constraints:
-                  BoxConstraints(maxWidth: MediaQuery.of(context).size.width),
-              child: sightCardVisited(context)),
-          child: sightCardVisited(context),
-          childWhenDragging: const SizedBox.shrink(),
-        );
-      },
-    );
-  }
-
-  Widget sightCardVisited(BuildContext context) {
-    return Material(
-      type: MaterialType.transparency,
-      child: InkWell(
-        onTap: () {
-          Navigator.push(
-              context,
-              MaterialPageRoute(
-                  builder: (context) => SightDetail(sight: sight)));
-        },
-        child: Container(
-          padding: EdgeInsets.only(left: 16, right: 16, top: 30, bottom: 30),
-          child: Column(
-            children: [
-              Stack(
-                children: [
-                  SizedBox(
-                    height: 30,
-                  ),
-                  Container(
-                    child: ClipRRect(
-                      borderRadius: const BorderRadius.only(
-                        topLeft: const Radius.circular(15),
-                        topRight: const Radius.circular(15),
-                      ),
-                      child: Image.network(
-                        '${sight.url}',
-                        height: double.infinity,
-                        width: double.infinity,
-                        fit: BoxFit.cover,
-                        loadingBuilder: (context, child, loadingProgress) {
-                          if (loadingProgress == null) return child;
-
-                          return Center(
-                            child: CircularProgressIndicator(
-                              value: loadingProgress.expectedTotalBytes != null
-                                  ? loadingProgress.cumulativeBytesLoaded /
-                                      loadingProgress.expectedTotalBytes
-                                  : null,
-                            ),
-                          );
-                        },
-                      ),
-                    ),
-                    height: Height,
-                    decoration: BoxDecoration(
-                      color: Theme.of(context).backgroundColor,
-                      borderRadius: const BorderRadius.only(
-                        topLeft: const Radius.circular(15),
-                        topRight: const Radius.circular(15),
-                      ),
-                    ),
-                  ),
-                  Positioned(
-                      top: 16,
-                      left: 16,
-                      child: Text(
-                        '${sight.titleType}',
-                        style: Theme.of(context).textTheme.bodyText2,
-                      )),
-                  Positioned(
-                    //!button share
-                    top: 1,
-                    right: 40,
-                    child: IconButton(
-                      icon: Icon(Icons.share, size: 28, color: iconColor),
-                      onPressed: () {
-                        print('button "share" pressed');
-                      },
-                    ),
-                  ),
-                  Positioned(
-                    //!button close
-                    top: 0,
-                    right: 0,
-                    child: IconButton(
-                      icon: Icon(Icons.close, size: 30, color: iconColor),
-                      onPressed: () {
-                        onDelete();
-                        print('fyui');
-                        sight.visited = !sight.visited;
-                      },
-                    ),
-                  ),
-                ],
-              ),
-              Container(
-                decoration: BoxDecoration(
-                  color: Theme.of(context).backgroundColor,
-                  borderRadius: const BorderRadius.only(
-                    bottomLeft: const Radius.circular(15),
-                    bottomRight: const Radius.circular(15),
-                  ),
-                ),
-                padding: const EdgeInsets.all(16),
-                width: double.infinity,
-                height: Height,
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      '${sight.name}',
-                      maxLines: 2,
-                      style: Theme.of(context).textTheme.headline5,
-                    ),
-                    SizedBox(
-                      height: 5,
-                    ),
-                    Text(
-                      'Цель достигнута 12 окт. 2020',
-                      style: Theme.of(context).textTheme.subtitle2,
-                    ),
-                    SizedBox(
-                      height: 5,
-                    ),
-                    Text(
-                      '${sight.workHours}',
-                      style: Theme.of(context).textTheme.subtitle2,
-                    ),
-                  ],
-                ),
-              ),
-            ],
-          ),
-        ),
       ),
     );
   }
